openPMD-api: One Series at a Time
Make sure that only one Series is open at a time in case `Init` is
called again.

In previous logic, at the moment of assignment of a new Series, the
old and the new `openPMD::Series`` were briefly active for a short
time. ADIOS1 does not support this and its anyway not what we
want to do here.

Discovered by Junmin.