Refactor AMReX IDs to Global ID
This PR exposes the AMReX AoS `.id` and `.cpu` to global
particle ID conversion for general usage in WarpX. The
generated particle ID is unique per particle in the sim.

Also replaces the old implementation with a well-documented
bitshift. The old implementation used C-style union type
punning, via access of the inactive union member, which is
technically undefined behavior (UB) in C++ (and is known to crash
on some platforms, e.g. MSVC). Although some popular compilers
implemented this C feature as an extension, it is never good
to rely on a specific interpretation of UB on a certain platform.