Docs: Summit Restructure
Restructures the instructions for Summit (OLCF).

This loads all software that might be needed to reduce the
"if then else" logic a user has to understand during install.
Everything is now required or optional, but everything is
compatible or simply ignored.

Uses new conventions for downloaded source and manually built
software to structure things in a simpler, scalable manner.