Travis CI: Wait Longer for Make
It looks like "make" run from our python test scripts does sometimes
take longer than 10 minutes and does not produce output on stdout
during that time (probably because we hide it).

This adds a 20min grace time for this (default: 10min) before ending
the job with a timeout.

Generally, tests on Travis are close to the 50min mark of total job
time, this patch does not change this :)