Docs: Ascent & Python Summit Updates (#1127)
* Docs: Ascent & Python Summit Updates

Updates the instructions on how to use Python with WarpX, e.g. for
libEnsemble runs on Summit.

Also updates Ascent location (compiles already with CMake).

* libEnsemble: backlink to Summit

* Summit libEnsemble: Missing steps

Document missing steps to be performed - thx Diana & Maxence!

* Summit FC: gfortran

* Summit: mpi4py

Not yet used with libEnsemble as we use it, but useful for yt
post-processing and future workflows.