Default: abort_on_out_of_gpu_memory = 1
Change the default input parameter from AMReX
`amrex.abort_on_out_of_gpu_memory` from false (`0`) to true (`1`).

We set this by default to avoid that users experience super-slow GPU
runs when exceeding GPU memory. In such a case, users should explicitly
set this to option.

In my optinion, this is only an intermediate solution since what we
actually want on out-of-GPU memory events should be:
- finish current simulation step and cause a load balance or
- trigger a checkpoint and shut down cleanly
  - then the user can manually restart with more resources

We want to address the opposite case, user under-utilizes a GPU, with
a warning for now.

Ref.:
- https://amrex-codes.github.io/amrex/docs_html/GPU.html#inputs-parameters