openPMD Mesh: Fix grid offset/spacing
Mesh output annotation for grid spacing and offset in openPMD was
annotated in inverse order. This made spacings appear off when
using a moving window and/or non-cubic cells.

This fixes it by writing ordered meta-data arrays in C-like order
(slowest to fastest varying index), which is also the order solely
used in openPMD 2.0.

Spotted by Rémi, thank you!

Co-authored-by: Remi Lehe <remi.lehe@normalesup.org>