openPMD: Generalize Mesh Position
This adds a small utility that returns the relative in-cell position of values
on each cell depending on an `amrex::MultiFab`'s `Box`'s `IndexType`.

This does not change the output of openPMD yet, since we spatially and
temporally center fields before output, but makes the routines
more general overall for future use.