MPI: Init or Thread Multiple
Simplify MPI init options to be independent of PSATD selection. Our
use of FFTD only uses the non-threaded FFTW routines.

Thread multiple is currently only needed for async I/O with AMReX
plotfiles.

Some MPI implementations allow to set the requested thread level with
`mpiexec` and/or pre-compiled variants of the lib and/or through an
environment variable, according to the standard. For example,
`MPICH_THREADLEVEL_DEFAULT` in MPICH (and derivatives have similar
names). This is done through an intricate upgrade/downgrade logic for
the provided argument (sec. 12.4.3). We print a warning if the
requested support level is not provided and a note if a stricter
support level is provided.