SYCL 2020/DPC++: C++17
Request C++17 in GNUmake for DPC++ builds. This is now required since
we use MKL headers that are in C++17 and since SYCL 2020 and later
rely on C++17 or newer.
(In fact, `dpcpp` beta07 and newer already defaulted to C++17.)

This updates the explicit GNUmake flag (request exactly one standard).

The CMake flags in WarpX are "request C++14 or newer" and AMReX
CMake files push those up to "request C++17 or newer" for DPC++.