AnyFFT: `AMReX_Config.H`
We reduce the number of command line arguments passed in the AMReX
build systems by placing defines in an `AMReX_Config.H` file that
is configured at AMReX build time. Files like `AMReX.H` will pull
it in as well.

In the case of `AnyFFT.H`, we need to make sure this is included
before we check the preprocessor defines.