Add "." to libwarpx Aliases
Add a "." to the suffix of the short-hand `libwarpx.<DIM>.<so|dll>`
libs. This unifies the suffix with the structure in executables and
makes it easier to find a "long" version of the library (not in this
PR).

Also fixes the suffix naming on Windows to ".dll", since this is
a true C/C++ library, not a Python module (aka not a shared library
following Python entry points and conventions).