FFTW: CMakeConfig Support
Modern FFTW versions support CMake as an install method, which
provides the legacy goodies (`.pc`) files and proper
`FFTW3(f)Config.cmake` files for dependency propagation.

Conda-forge uses the CMake-based method to deploy. Checking this,
I realized the `.pc` files add unconditionally `libm` to the linker,
which is wrong on Windows and fails the builds.

Thus, forward looking we would search first for CMake based installs
of FFTW and in case this search is not successful, we continue to
search for the old Autotools logic and `.pc` files, which are still
routinely generated by some package managers.
But, due to an issue in FFTW3 3.3.9 on Linux/macOS with Autotools
installs, we make `Config.cmake` file based search still opt-in for
now.