Parser: no jnf for Apple
At least on AppleClang and also seen in flang source-code, we
just call the double variant of `jn` instead of using a specialized
call to `jnf`.

This might be less efficient than needed with Intel on Apple, but
I don't have a system to test this on at the moment and this is thus
a save bet for compile stability.