paraview.pmd: write on close (#1860)
Initially we wrote those on open. The problem was that writing this
file depends on an existing directory structure, which is first
created on earlierst on the first `flush()` and in case of extensive
ADIOS buffering actually latest on `close()`.

Since in backtransformed diagnostics (BTD) this flush happens very
late. The directory did not exist and thus the creation of the
helper file early on did fail (silently).