Fix LoadBalance + MR: Rebuild Particle Masks (#1904)
In MR simulations, particle masks are used to determine if a
fine-patch particle is in the field gather and/or current deposition
buffer regions near the coarse/fine boundaries. This is needed
because some particles deposit to / gather from the coarse level,
since they are near the level boundaries.

On regriding during load-balancing, we can avoid to communicate the
`MultiFab`s for those masks (gather and deposit) if we locally
re-build the mask in the new distribution mapping. We forgot to
trigger the rebuild, which caused illegal memory accesses in the
particle evolve (partition) down the road.