openPMD: Expose ADIOS2 Operators (Compressors)
Expose ADIOS2 Operators (at the moment: one operator max) in WarpX
inputs. This allows the user to select ZFP or Blosc compression
without changing the input file.

We can later on define short-hand notions for this. For now we give
the parameters control out for the expert user and document good
examples.