Azure: Always with EB
Always compile with EB in Azure.
Since EB shall be an opt-in, non-exclusive option we can also
just compile with it to make sure we cover all geometries & cases.

Also compiles EB with all compilers in GitHub actions, besides
one entry (minimal build) to make sure we can continue to build
without it.