Fix: CUDA -O3 with CMake
`-O3` was only set for CUDA if we ran CMake at least twice. This was
an intialization logic bug and should be set from the beginning, but
we initialized CUDA too late / applied the transform too early.

We now change the default build type to `Release` to avoid changing
defaults and avoid further surprises.