WarpX::Version: Not Empty
Make sure we do not return an empty string in `WarpX::Version()`.
In some situations, e.g., in CI/run_test.sh, the macro for the GIT
version in GNUmake is set but empty.

Since empty versions are problematic for HDF5 attributes and
confusing anyway, we return a proper "Unknown" in such a situation
now, too.