GetPosition 2D3V: y=0
In 2D3V, we currently assigned the y position a signaling NaN in
`GetPosition`, so that it cannot be used accidentially in functions.

When we call parser functions, this value is copied even if it is not
used, raising the signaling NaN.

Rethinking the strategy, we can write more general inputs files if
we instead write a sensible default for `y` in 2D3V: `y=0`. So this
PR changes the default, which also fixes the issue.