Docs: Summit umask (Permissions)
After the update to RHEL8, Summit changed the default
permissions of newly created files and directories (for the
worse).

This has been reported as OLCFHELP-3442 but we face some
resistance from support to triage this properly at the moment.

Since we need to continue to keep working, we change the `umask`
(aka defaults for new files & dirs) manually so that group
members of the same project can read files and access dirs.

For files and dirs created since this update and not yet using
this `umask`, please use the following fix.
```
find . -type -d -exec chmod a+rx {} \;
find . -type -f -exec chmod a+r {} \;
```

Replace `.` (current directory) with another path if needed.