Python Install Env Vars: All CAPS
On Windows, environment variable names are case insensitive and often
mapped to all CAPS spelling.

- https://stackoverflow.com/questions/7797269/python-environment-case-senstivity-os-environ
- https://en.wikipedia.org/wiki/Environment_variable#Syntax

This leads to an inconsistency between Unix and Windows like systems
if we use case sensitive environment variable names for install
control in Python.

Luckily, it is quite common to use all-CAPS variables in UNIX for
environment controls as well, so we will make a breaking change to
unify on that. Existing usage will see a warning of an unused
variable with `python3 -m pip install -v .`