Fix bug and pass TemperatureInit by reference
- Fixed bug where the parser in TemperatureInit was going out of scope
before we were done using the executor. This has been fixed by making
the parser a member of the PlasmaInjector class so it won't be
prematurely garbage collected. This should be temporary until the
refactor of TemperatureInit
- TemperatureInit is no longer a union, and is instead just a struct
whose first member indicates the type
- The PlasmaInjector method that reads the temperature parameters is now
void. It takes a TemperatureInit object by reference and does the real
initialization.