Restructure how temperature data are stored
- TemperatureInit renamed to TemperatureProperties
- Temperature data is stored in an instance of TemperatureProperties,
a smart pointer to which is owned by PlasmaInjector.
- TemperatureProperties is initialized with a reference to parameters,
and owns the temperature function parser, if used.
- GetTemperature is a functor that takes three doubles (x, y, and z
coordinates) and returns the initial temperature at that location. This
is either a constant or a parser evaluated at (x,y,z), depending on
input parameters.
- InjectorMomentum takes a const reference to GetTemperature in its
constructor when needed and stores a copy. The compiled parser will
remain in scope because it is owned by TemperatureProperties, which is
in turn owned by PlasmaInjector.