Bilinear Filter: Fix Warning (Vector)
GCC warns in some versions on the access here, since it cannot figure
out the runtime size of the vectors.

We can just replace `[]` with `.at()` to use a range-checked access
that throws an exception instead of a segfault/UB when accessed
out-of-bounds. In the preparation of the compute stencils, we don't
have hot loops that benefit from non-range checked access, so that
cost for improved safety is fine.