CMake: 3.18+
With the C++17 switch, we required CMake 3.17+ since that one
introduced the `cuda_std_17` target compile feature.

It turns out that one of the many CUDA improvements in CMake 3.18+
is also to fix that feature for good, so we bump our requirement in
CMake. Since CMake is easy to install, it's easier to require a clean
newer version than working around a broken old one.

Spotted first by Phil on AWS instances, thx!