Fix: Move MPI Thread Level Check
Move the MPI thread level check (requested vs. provided) into `InitData()`.
This is needed, since we will access the warning
logger, which itself needs to be accessed via a WarpX instance.

This is also cleaner than just moving this behind the constructor
in `main.cpp`, as we have less functions to call around the
`WarpX` object usage.