Clean Pip Staging Directory
The distutils staging (`build` -> `_tmppythonbuild`) directory that
pip uses to collect build artifacts is not by default cleaned between
multiple `pip` runs.

This can be confusing when we recompile, because old `libwarpx*` files
can be still in it that are not present in our own `build/lib/`.

This cleans that staging directory before builds now.
It also sets the whole `build_base` so that no artifact lands in the
default, which was `build/`. So far, `bdist.linux-x86_64/` still was
out-of-tree.