Fix SyncCurrent and SyncRho for multi-levels
Previously these two function did not work when max_level > 1.  For example,
when max_level = 2, the coarse patch data from level 2 were coarsened and
added to the fine patch on level 1. However, the coarse patch on level 1 did
not have any contribution from level 2 as it should have, because the
coarsening was performed before the fine patch on level 1 received the level
2 contribution.  The order has been fixed in this commit.