DD fusion anneutronic branch (#3789)
* Allow DD fusion products to be tritium + protium

For deuterium-deuterium (DD) fusion there are two branches:

           | He-3 + n (neutronic branch, 50% probability)
 d + d -> <
           | t + p    (anneutronic branch, 50% probability)

Currently in Warpx anneutronic DD fusion only works when the species of
the fusion products are tritium (or hydrogen3) and proton. If protium
(or hydrogen1) are specified instead of proton, the simulation aborts
with the following error message:

amrex::Abort::0::ERROR: Product species of proton-boron fusion must be of type helium3 and neutron, or tritium and proton !!!

The proposed diff allows anneutronic DD fusion to produce either protium
/ hydrogen1 or proton. It also corrects the error message.

* Shows that DD fusion cannot produce protium

Input deck that demonstrates how anneutronic DD fusion in the
development branch cannot produce protium, only proton.

Some options of the anneutronic DD branch:
1) tritium + proton (currently the only option)
2) tritium + protium (added in this branch)
3) triton + proton (another alternative)

Option 1) seems inconsistent (why include the electron mass in one
fusion product species but not in the other?). Option 2) is enabled
by the diff in this branch. Option 3) would require adding a triton
species (and deuteron and helion species for consistency).

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Remove test file

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Remi Lehe <remi.lehe@normalesup.org>