`geometry.is_periodic`: Only Warn (#3974)
We want to be able to directly run a simulation from the created
`warpx_used_inputs` file of a previous simulation. Since we use
`geometry.is_periodic` internally to communicate boundary
conditions to AMReX solvers and communication routines, we cannot
fail if it is present but instead should only warn.