Doc: HPC Python Separate
Currently, due to global variables used in AMReX to store implicit
state instead of using handlers, we cannot build shared libraries
for Python (pyAMReX + pyWarpX + pyImpactX) with a *static*
`libamrex_Nd.a`. Thus, we automatically build a shared AMReX
library in our superbuild the moment that we need Python bindings.

That can confuse people that just use the C++ application with
AMReX inputs files, which then also depends on this shared AMReX
library (which also needs to be copied around, for instance).

This update the HPC docs to just build App and bindings twice
(twice the time spent), to work around this current limitation
of AMReX.