Disable ccache by default
WarpX autodetects `ccache` and uses it, and there's nothing you can do
about it. This PR disables it by default, and lets developers enable it
through `-DWarpX_CCACHE:BOOL=ON`.

The reason for this is mostly related to compiler wrappers, of which
there are many...

If `g++` is a compiler wrapper, then `ccache g++` cannot see the
effective flags passed to underlying real `g++`. That leads eventually
to false positive cache hits, which is a pain to debug.

For compiler wrappers you want `g++` the wrapper to invoke the call to
`ccache <real g++ invocation>` to fix that, which is for example how
Spack handle it.

Further, if you use spack with ccache enabled, the defaults of WarpX
cause `ccache` to be invoked twice (inner & outer), doubling the cache
requirements.

Finally, compiler wrapper that handle ccache themselves may set further
ccache options / flags that WarpX does not set, such as disabling
hashing of the build dir -- w/o that option cache may be useless.