Fix: heFFTe on Perlmutter GPU
heFFTe usage with NVCC 12.2 on Perlmutter fails to compile WarpX
once its public header is included, due to unguarded include of
a CPU AVX intrincics header in its stock backend.

We do not need a fast stock CPU backend if we compile for CUDA,
so we patch it out now to avoid the issue. A future release
of heFFTe will probably ship our fix for this, until then
we post-install patch heFFTe on Perlmutter for GPU builds.