warpx.roundrobin_sfc: A runtime parameter to control distribution mapping (#4909)
* warpx.roundrobin_sfc: A runtime parameter to control distribution mapping

The default is false. If it's true, AMReX's RRSFS strategy will be used to
override the default SFC strategy used by amrex::AmrCore. The motivation for
this is that this might mitigate the load imbalance issue during
initialization by avoiding putting neighboring boxes on the same process.

* Update parameters.rst

* Add control in PICMI

* Remove WarpX:roundrobin_sfc

* add const

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>