GatherAndPush: Use CTO ParallelFor
The GPU kernel in GatherAndPush contains Parsers for computing external E
and B fields. It's known that even if the Parsers are not used at run time,
the occupancy on AMD GPUs are still affected. This is probably reason why
this kernel is 50% slower on MI200 than on A100. Using CTO ParallelFor
should improve the performance.