HPC: `SRUN_CPUS_PER_TASK` to `SLURM_`
This environment variable was used for Perlmutter when
`--cpus-per-task=N` did not work yet. It was copied around to
other templates.

These days, `--cpus-per-task` should work and the name of the
env variable was renamed in SLURM to `SLURM_CPUS_PER_TASK`.
https://slurm.schedmd.com/sbatch.html#OPT_SLURM_CPUS_PER_TASK

Thanks to NERSC engineers for reporting this update!